rename keydown class to shortcuts
